LCD tests: introduce checkPartialScreenshot()
The checkPartialScreenshot function compares the contents of the LCD buffer
with a reference image of a possibly different size than the LCD buffer. The
part of the image that due to different dimensions is not present in either the
LCD buffer or the reference picture is assumed to have the background color
(i.e. unset pixels). This allows to reduce the number of reference images for
tests for which the LCD size doesn't matter.

The checkScreenshot function is now just a wrapper around checkPartialScreenshot.

The implementation uses a different approach than the one used so far. Instead
of writing the rendered picture to a file in /tmp/ and comparing contents of
the file with the reference file, the images are compared in memory and files
are only written when a mismatch occurs. In such a case, additionally to the
rendered LCD buffer a difference file is generated.