LCD: fix TINSIZE on non-Taranis ARM (e.g. 9XRPRO)
The bug was introduced in 059ea49f on the wake of fixing #2183.  There was a
change in TINSIZE glyph layout and the necessary follow-up was only in
gui/Taranis/lcd.cpp and thus implementation in gui/9X/lcd.cpp ceased to work.

It wouldn't have happened if we had less code duplication (#2315).

Partially fixes outdezNAtt.tinsize test for non-Taranis ARM builds.