Issue #4224. Allow pasting models into and across categories.
To do this, we need both the current category and model index
passed into doPaste(), which then decides if the user is pasting
into a category based on whether or not modelIdx == -1.

Doing a category paste just means we stub out a ModelData object,
insert it into the models vector then use the existing code to
paste over it.

Also, the code now allows users to copy/paste models across
categories.