Issue #4368. Re-work commit ea44875 to correct some problems.
This is a reworked version of the  code in commit ea44875. I think it
is  now cleaner and correct for both the category and non-category case.

While iterating thru the models vector, we either add a line of text to
the in-memory models.txt or add a line to the sortedModels map.  I got
rid of rhe map trying to handle both cases. Now we do one or the other,
swithcing both the format of the line and the "needs sorting" code path
based on the  "HasModelCategories" capability of the firmare.