Issue #4375, fix pasting multiple items in Companion.
When pasting multiple items in Companion, we have to follow the UI to
determine the next row to paste into rather than just increasing model
index. Also, with a category capable firmware, if we paste more than
one item "off the end" of the category, the extra items should go end
up in the same category.