Add support for retrieving model ID over CRSF
Add 2 frames types:

- REQUEST_MODEL_CONFIG_ID sent from the TX to the radio to request
the model configuration. Empty.
- MODEL_CONFIG_ID sent from the radio to the TX. Contains the number
of bytes of the model ID followed by the ID itself and the
null-terminated model name.

For retrieving the model ID, the TX sends REQUEST_MODEL_CONFIG_ID,
which makes the radio send MODEL_CONFIG_ID rather than CHANNELS_ID
in the next cycle. This allows the TX to only ask for the model ID
when it needs it without loosing a control cycle periodically.