Improve latency for generated pulses
We run the mixer 2ms before setupPules and setupPules 2ms before
actually starting to send the pulses. The last 2ms are very pessimistic.
This commit changes that time to a more realistic but still more
than comfortable 150µs.

In my own tests the setup of pulses was always < 70 µs but be still
conservative here.

Fix multi sync rate to be always 7ms when no sync signal is present and
adjust mixer interval to sync rate.