[USB] Power up immediately when booting with USB connected
This allows jumping out of ST's bootloader without manual user
intervention when the radio is connected to the USB host while
it's powered off.

Note that the radio doesn't power on as soon as it's connected, since
the hardware is wired to boot the MCU into bootloader mode. However,
the host can detect the USB connection and issue a DFU_DNLOAD to
turn on the radio.