[MSC] Add a file in the virtual Taranis drive with radio information
Add a new readonly file called RADIO.TXT to the root directory.
This file contains the unique ID of the CPU as well as a DFU
serial using the same algorithm as ST's DFU bootloader.

This way we can keep track of an specific radio when it's connected
in MSC mode.