[build] Fix building with semihosting enabled
Remove -DSEMIHOSTING from the bootloader CFLAGS unconditionally.

-DSEMIHOSTING is passed to the bootloader by the parent project when
SEMIHOSTING=ON, however the bootloader wouldn't fit on flash with
semihosting enabled.

If we leave the -DSEMIHOSTING flag, we would get a compilation error
on non-horus radios (since -DDEBUG is removed) and a link time error
on horus (since the bootloader won't fit in its allocated flash
region).